requirements.txt
Random-Word==1.0.4
Image==1.5.27
psycopg2-binary
httplib2